moved TestRaftCluster to test_utils and refactored some tests to use the test raft cluster. everything working. flaky tests seem to be fixed.
